完成了SpringCache缓存，这里记一个坑，不要去设置Redis序列化值的方式，否则反序列化时侯跟Jackson冲突会导致失败

+++ src/main/java/com/zijun/reggie/common/EmployeeLoginCheckInterceptor.java
@@ -28,7 +28,7 @@ public class EmployeeLoginCheckInterceptor extends HandlerInterceptorAdapter {
   @Override
   public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
 
-    Integer employeeId = (Integer) redisTemplate.opsForValue().get("employee");
+    Long employeeId = (Long) redisTemplate.opsForValue().get("employee");
     String json = "";
     if (employeeId != null) {
       LambdaQueryWrapper<Employee> wrapper = new LambdaQueryWrapper<>();
